third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGPatternElement-dom-patternUnits-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'patternUnits' attribute of the SVGPatternElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS patternElement.getAttribute('patternUnits') is null
PASS patternElement.getAttribute('patternUnits') is "userSpaceOnUse"
PASS successfullyParsed is true

TEST COMPLETE
